+++ app/Jobs/ProcessSignRequest.php
@@ -2,6 +2,7 @@
 
 namespace App\Jobs;
 
+use App\Jobs\ProcessDeleteFile;
 use App\Notifications\SignRequest\Requester;
 use App\Notifications\SignRequest\Signer;
 use Illuminate\Bus\Queueable;
@@ -33,7 +34,10 @@ class ProcessSignRequest implements ShouldQueue
      */
     public function handle()
     {
+        $encrypted_filename = hash('sha256', $this->data['email'] . $this->data['filename'] . $this->data['id']);
+
         Notification::route('mail', $this->data['email'])->notify(new Requester($this->data));
         Notification::route('mail', $this->data['signers'][0]['email'])->notify(new Signer($this->data));
+        dispatch(new ProcessDeleteFIle($encrypted_filename))->delay(now()->addMinutes(720));
     }
 }